renamed 2 folders to utilize proper naming conventions
